+++ src/webview/icons/index.ts
@@ -13,6 +13,13 @@ import boltSvg from "./bolt.svg";
 import wrenchSvg from "./wrench.svg";
 import codeMergeSvg from "./code-merge.svg";
 import flaskSvg from "./flask.svg";
+import decisionSvg from "./decision.svg";
+import messageSvg from "./message.svg";
+import gateSvg from "./gate.svg";
+import spikeSvg from "./spike.svg";
+import storySvg from "./story.svg";
+import milestoneSvg from "./milestone.svg";
+import eventSvg from "./event.svg";
 
 // UI icons
 import userSvg from "./user.svg";
@@ -29,6 +36,13 @@ export const icons = {
   chore: wrenchSvg,
   "merge-request": codeMergeSvg,
   molecule: flaskSvg,
+  decision: decisionSvg,
+  message: messageSvg,
+  gate: gateSvg,
+  spike: spikeSvg,
+  story: storySvg,
+  milestone: milestoneSvg,
+  event: eventSvg,
   // UI
   user: userSvg,
   tag: tagSvg,